requirements.txt
pandas
python-dotenv
openai
tabulate
pytest
